Rename each SubsystemParsingTestCase by prefixing with respective subsystem name
