+++ batch/extension/src/test/java/org/wildfly/extension/batch/BatchSubsystemParsingTestCase.java
@@ -1,6 +1,6 @@
 /*
  * JBoss, Home of Professional Open Source.
- * Copyright 2015, Red Hat, Inc., and individual contributors
+ * Copyright 2013, Red Hat, Inc., and individual contributors
  * as indicated by the @author tags. See the copyright.txt file in the
  * distribution for a full listing of individual contributors.
  *
@@ -20,19 +20,18 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.wildfly.extension.batch.jberet;
+package org.wildfly.extension.batch;
 
 import java.io.IOException;
 
-import org.jboss.as.subsystem.test.AdditionalInitialization;
 import org.junit.Test;
 
 /**
  * Basic subsystem test. Tests parsing various batch configurations
  */
-public class SubsystemParsingTestCase extends AbstractBatchTestCase {
+public class BatchSubsystemParsingTestCase extends AbstractBatchTestCase {
 
-    public SubsystemParsingTestCase() {
+    public BatchSubsystemParsingTestCase() {
         super(BatchSubsystemDefinition.NAME, new BatchSubsystemExtension());
     }
 
@@ -43,13 +42,13 @@ public class SubsystemParsingTestCase extends AbstractBatchTestCase {
 
     @Override
     protected String getSubsystemXsdPath() throws Exception {
-        return "schema/wildfly-batch-jberet_1_0.xsd";
+        return "schema/wildfly-batch_1_0.xsd";
     }
 
     @Override
     protected String[] getSubsystemTemplatePaths() throws IOException {
         return new String[] {
-            "/subsystem-templates/batch-jberet.xml"
+            "/subsystem-templates/batch.xml"
         };
     }
 
@@ -66,10 +65,6 @@ public class SubsystemParsingTestCase extends AbstractBatchTestCase {
     @Test
     public void testJdbcSubsystem() throws Exception {
         standardSubsystemTest("/jdbc-default-subsystem.xml");
-    }
-
-    @Override
-    protected AdditionalInitialization createAdditionalInitialization() {
-        return AdditionalInitialization.withCapabilities("org.wildfly.data-source.ExampleDS");
+        standardSubsystemTest("/jdbc-subsystem.xml");
     }
 }
